chore: move css helper to plugin-typed-css-modules
