+++ packages/plugin-typed-css-modules/src/loader.ts
@@ -12,13 +12,19 @@
 
 import fs from 'node:fs';
 import path from 'node:path';
-import {
-  type CssLoaderModules,
-  type Rspack,
-  isCssModules,
-  isInNodeModules,
-} from '@rsbuild/shared';
+import { NODE_MODULES_REGEX, type Rspack } from '@rsbuild/shared';
 import LineDiff from '../compiled/line-diff/index.js';
+
+export type CssLoaderModules =
+  | boolean
+  | string
+  | {
+      auto: boolean | RegExp | ((filename: string) => boolean);
+    };
+
+export const isInNodeModules = (path: string) => NODE_MODULES_REGEX.test(path);
+
+export const CSS_MODULES_REGEX = /\.module\.\w+$/i;
 
 const bannerMessage =
   '// This file is automatically generated.\n// Please do not change this file!';
@@ -30,6 +36,32 @@ const getNoDeclarationFileError = ({ filename }: { filename: string }) =>
   new Error(
     `Generated type declaration does not exist. Run webpack and commit the type declaration for '${filename}'`,
   );
+
+export const isCssModules = (filename: string, modules: CssLoaderModules) => {
+  if (typeof modules === 'boolean') {
+    return modules;
+  }
+
+  // Same as the `mode` option
+  // https://github.com/webpack-contrib/css-loader?tab=readme-ov-file#mode
+  if (typeof modules === 'string') {
+    // CSS Modules will be disabled if mode is 'global'
+    return modules !== 'global';
+  }
+
+  const { auto } = modules;
+
+  if (typeof auto === 'boolean') {
+    return auto && CSS_MODULES_REGEX.test(filename);
+  }
+  if (auto instanceof RegExp) {
+    return auto.test(filename);
+  }
+  if (typeof auto === 'function') {
+    return auto(filename);
+  }
+  return true;
+};
 
 const getTypeMismatchError = ({
   filename,
