+++ packages/shared/src/constants.ts
@@ -65,7 +65,6 @@ export const SVG_REGEX = /\.svg$/;
 export const CSS_REGEX = /\.css$/;
 export const LESS_REGEX = /\.less$/;
 export const SASS_REGEX = /\.s(?:a|c)ss$/;
-export const CSS_MODULES_REGEX = /\.module\.\w+$/i;
 export const NODE_MODULES_REGEX = /[\\/]node_modules[\\/]/;
 
 export const TS_CONFIG_FILE = 'tsconfig.json';
